feat: Rename socket folder and event #61
