+++ src/Server/event/error.cpp
@@ -1,0 +1,7 @@
+#include "Server.hpp"
+
+void Server::serverErrorEvent(struct kevent *current_event)
+{
+  std::cout << "server socket error" << std::endl;
+  disconnect_socket(current_event->ident);
+}
